# backend/docs/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: CozyVTT API
  description: |
    RESTful API for CozyVTT — a self-hosted, browser-based Virtual Tabletop platform.

    ## Authentication
    All endpoints (except `/health`, `/api/setup/*`, and public auth routes) require authentication via session cookies.
    Login via `/api/auth/login` to obtain a session cookie.

    ## MFA Flow
    If a user has MFA enabled, `/api/auth/login` returns `{ mfaRequired: true }` instead of a full session.
    The client must then call `/api/auth/mfa/verify-login` with a TOTP code or backup code to complete authentication.

    ## Authorization
    - **Platform roles**: `ADMIN` (full system access), `USER` (campaign access)
    - **Campaign roles**: `DM` (full campaign control), `PLAYER` (limited interaction), `SPECTATOR` (read-only)

    ## Rate Limiting
    - Global API limit: 300 requests per minute per IP (all `/api/*` routes)
    - Authentication endpoints: 5 requests per 15 minutes
    - MFA setup verification: 5 attempts per 15 minutes
    - MFA login verification: 5 attempts per 15 minutes (separate limit)
    - Asset uploads: 30 requests per minute per user (configurable via `ASSET_UPLOAD_RATE_LIMIT`)
    - Chat messages: 5 per minute (WebSocket)
    - Dice rolls: 30 per minute (WebSocket)
    - Token movement: 60 per second (WebSocket, throttled)

    Rate-limited responses return HTTP 429 with `RateLimit-*` standard headers.

    ## Forced password changes
...

    ## Real-time Events
...

  version: 1.2.1
...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: List all users
      description: Get list of all users (Admin only)
      operationId: listUsers
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/campaigns:
    get:
      tags:
        - Campaigns
      summary: List user campaigns
      description: Get all campaigns the user is a member of
      operationId: listCampaigns
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaigns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'

    post:
      tags:
        - Campaigns
      summary: Create campaign
      description: |
        Create a new campaign (creator automatically becomes DM). The optional
        `gameSystem` field is preferred but not required; it can be set later.
      operationId: createCampaign
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                gameSystem:
                  type: string
                  enum: [DND_5E, PATHFINDER_1E, PATHFINDER_2E, SHADOWRUN_6E, CALL_OF_CTHULHU_7E]
                  nullable: true
                  description: Optional game system for the campaign
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  campaign:
                    $ref: '#/components/schemas/Campaign'
...
  /api/invitations:
    get:
      tags:
        - Invitations
      summary: List my pending invitations
      description: |
        Returns the authenticated user's pending campaign invitations (status
        `PENDING`), ordered newest first, with the inviting campaign's summary
        info included so the UI can render an invitation card without
        additional fetches.
      operationId: listMyInvitations
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Pending invitations.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    campaignId: { type: string, format: uuid }
                    userId: { type: string, format: uuid }
                    status:
                      type: string
                      enum: [PENDING, ACCEPTED, DECLINED, EXPIRED]
                    createdAt: { type: string, format: date-time }
                    expiresAt: { type: string, format: date-time, nullable: true }
                    campaign:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        name: { type: string }
                        description: { type: string, nullable: true }
                        gameSystem: { type: string, nullable: true }
                        status:
                          type: string
                          enum: [PREPARATION, ACTIVE, PAUSED, COMPLETED, ARCHIVED]
                        owner:
                          type: object
                          properties:
                            displayName: { type: string }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/characters:
    get:
      tags:
        - Characters
      summary: List user characters
      description: Get all characters owned by authenticated user
      operationId: listCharacters
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of characters
          content:
            application/json:
              schema:
                type: object
                properties:
                  characters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Character'

    post:
      tags:
        - Characters
      summary: Create character
      description: |
        Create a new character.
        - If `campaignId` is provided and `gameSystem` is not specified, the
          character auto-inherits the campaign's `gameSystem`.
        - If `gameSystem` is provided, character data is validated against that
          system's Zod schema.
        - If `gameSystem` is null, no validation occurs (flexible character).
      operationId: createCharacter
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                gameSystem:
                  type: string
                  enum: [DND_5E, PATHFINDER_1E, PATHFINDER_2E, SHADOWRUN_6E, CALL_OF_CTHULHU_7E]
                  nullable: true
                  description: Optional game system (auto-inherited from campaign if campaignId provided)
                campaignId:
                  type: string
                  format: uuid
                  nullable: true
                  description: Optional campaign to assign character to immediately
                data:
                  type: object
                  description: Flexible JSON data for character stats (validated if gameSystem is set)
                tokenImageUrl:
                  type: string
      responses:
        '201':
          description: Character created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  character:
                    $ref: '#/components/schemas/Character'
        '400':
          description: Validation error (invalid gameSystem or character data doesn't match schema)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Validation Error
                  message:
                    type: string
                  validationErrors:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        message:
                          type: string
                        code:
                          type: string
...
  /api/assets:
    get:
      tags:
        - Assets
      summary: List assets
      description: List assets with filtering and pagination
      operationId: listAssets
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [MAP, TOKEN, AUDIO, AVATAR]
        - name: scope
          in: query
          schema:
            type: string
            enum: [GLOBAL, CAMPAIGN]
        - name: campaignId
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
...
  /api/campaigns/{campaignId}/maps:
    get:
      tags:
        - Maps
      summary: List campaign maps
      description: Get all maps for a campaign (campaign member)
      operationId: listMaps
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      responses:
        '200':
          description: List of maps
          content:
            application/json:
              schema:
                type: object
                properties:
                  maps:
                    type: array
                    items:
                      $ref: '#/components/schemas/Map'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Maps
      summary: Create map
      description: Create a new map for the campaign (DM only)
      operationId: createMap
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - imageUrl
                - width
                - height
              properties:
                name:
                  type: string
                imageUrl:
                  type: string
                width:
                  type: integer
                  minimum: 1
                height:
                  type: integer
                  minimum: 1
                gridSize:
                  type: integer
                  minimum: 1
                  default: 50
                spiritLayerUrl:
                  type: string
      responses:
        '201':
          description: Map created
          content:
            application/json:
              schema:
                type: object
                properties:
                  map:
                    $ref: '#/components/schemas/Map'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/campaigns/{campaignId}/creatures:
    get:
      tags: [Creatures]
      summary: List creature templates
      description: |
        Returns SRD (global) + campaign-specific custom creature templates.
        Supports search, source filter, challenge rating filter, and pagination.
        Any campaign member can list.
      operationId: listCreatures
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: search
          in: query
          schema:
            type: string
          description: Filter by name (case-insensitive partial match)
        - name: source
          in: query
          schema:
            type: string
            enum: [srd, custom]
          description: Filter by source
        - name: cr
          in: query
          schema:
            type: string
          description: Filter by challenge rating
        - name: gameSystem
          in: query
          schema:
            type: string
          description: Filter by game system
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated creature list
          content:
            application/json:
              schema:
                type: object
                properties:
                  creatures:
                    type: array
                    items:
                      $ref: '#/components/schemas/CreatureTemplate'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Creatures]
      summary: Create custom creature
      description: Create a new campaign-specific custom creature template (DM only).
      operationId: createCreature
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - statBlock
              properties:
                name:
                  type: string
                statBlock:
                  type: object
                gameSystem:
                  type: string
                challengeRating:
                  type: string
                creatureType:
                  type: string
                alignment:
                  type: string
                imageUrl:
                  type: string
                size:
                  type: object
                  properties:
                    width:
                      type: integer
                    height:
                      type: integer
                disposition:
                  type: string
                  enum: [hostile, friendly, neutral]
                displayMode:
                  type: string
                  enum: [pog, top-down, full-art]
      responses:
        '201':
          description: Creature template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatureTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/campaigns/{campaignId}/token-templates:
    get:
      tags: [Token Templates]
      summary: List token templates
      description: Get all token templates for a campaign. Supports search and type filtering.
      operationId: listTokenTemplates
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [npc, player, object]
      responses:
        '200':
          description: List of token templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/TokenTemplate'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Token Templates]
      summary: Create token template
      description: Create a new token template (DM only).
      operationId: createTokenTemplate
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenTemplateInput'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/config:
    get:
      tags: [Admin]
      summary: Get server configuration
      description: Returns read-only server configuration — upload size limits, session timeouts, and SMTP status (no credentials exposed).
      operationId: adminGetConfig
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Server configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadLimits:
                    type: object
                    properties:
                      MAP:
                        type: integer
                        description: Max file size in bytes
                      TOKEN:
                        type: integer
                      AUDIO:
                        type: integer
                      AVATAR:
                        type: integer
                  sessionTimeoutMs:
                    type: integer
                  rememberMeTimeoutMs:
                    type: integer
                  smtp:
                    type: object
                    properties:
                      configured:
                        type: boolean
                      host:
                        type: string
                        nullable: true
                      port:
                        type: integer
                      user:
                        type: string
                        nullable: true
                      secure:
                        type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/users:
    post:
      tags: [Admin]
      summary: Create a new user
      description: |
        Creates a user with a generated temporary password. The account is flagged
        `mustChangePassword`, so that password can only be used to set a real one — every
        other endpoint returns 403 `PASSWORD_CHANGE_REQUIRED` until it is replaced.

        When SMTP is configured the credentials are emailed and `temporaryPassword` is
        **withheld** from the response; it is returned only when the email could not be
        sent, so the admin never holds a working credential unnecessarily. To create an
        account without any password at all, use `/api/admin/users/invite`.
      operationId: adminCreateUser
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                displayName:
                  type: string
                  maxLength: 50
                platformRole:
                  type: string
                  enum: [USER, ADMIN]
                  default: USER
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  emailSent:
                    type: boolean
                    description: Whether the welcome email with the credentials was delivered
                  temporaryPassword:
                    type: string
                    description: >-
                      One-time temporary password. Present ONLY when emailSent is false —
                      share it securely; it is not retrievable afterwards.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already in use
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cozyvtt.sid
      description: Session cookie obtained via /api/auth/login
  parameters:
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Campaign UUID
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        platformRole:
          type: string
          enum: [ADMIN, USER]
        mfaEnabled:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        gameSystem:
          type: string
          enum: [DND_5E, PATHFINDER_1E, PATHFINDER_2E, SHADOWRUN_6E, CALL_OF_CTHULHU_7E]
          nullable: true
          description: Primary game system for this campaign (optional)
        ownerId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PREPARATION, ACTIVE, PAUSED, COMPLETED, ARCHIVED]
          default: PREPARATION
        currentMapId:
          type: string
          format: uuid
          nullable: true
        vibeSettings:
          type: object
          description: Vibe periods with hues, filters, and audio
        currentVibe:
          type: string
          nullable: true
          description: Current time period name
        spiritLayerEnabled:
          type: boolean
          default: false
        spiritLayerStyle:
          type: string
          default: wispy
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastPlayedAt:
          type: string
          format: date-time
          nullable: true
...
    Character:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        gameSystem:
          type: string
          enum: [DND_5E, PATHFINDER_1E, PATHFINDER_2E, SHADOWRUN_6E, CALL_OF_CTHULHU_7E]
          nullable: true
          description: Game system for this character (determines data structure)
        data:
          type: object
          description: Flexible JSON data for character stats (structure depends on gameSystem)
        tokenImageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [MAP, TOKEN, AUDIO, AVATAR]
        scope:
          type: string
          enum: [GLOBAL, CAMPAIGN]
        uploadedById:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
          nullable: true
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        filePath:
          type: string
        thumbnailPath:
          type: string
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input or validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BadRequest
            message: Required field "name" is missing.
    Unauthorized:
      description: Unauthorized - Not authenticated or MFA pending.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Authentication required.
    Forbidden:
      description: Forbidden - Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: You do not have permission to perform this action.
...
    NotFound:
      description: Not Found - Resource does not exist or you cannot see it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Campaign not found.
...
    ServerError:
      description: Internal Server Error - Something went wrong on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: InternalServerError
            message: An unexpected error occurred. Please try again later.